requirements.txt
Pillow==8.1.2
pytube==10.4.1
requests==2.25.1